# api/openapi.yaml
openapi: 3.0.3
info:
  title: Convia API
  version: 0.1.0
  description: |
    Convia is a real-time communication platform. This document is the contract
    for its public HTTP API and for the operational endpoints that surround it.

    Media transport is an internal implementation detail. No field, schema, or
    error in this contract may expose a media provider concept.

    The transport rules that every endpoint follows are described in
    `docs/api-conventions.md`. The compatibility rules that govern changes to
    this document are described in `docs/api-compatibility.md`.
  license:
    name: PolyForm Noncommercial License 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0
servers:
  - url: http://localhost:8080
    description: Local development instance.
...
paths:
...
  /v1/applications:
    post:
      operationId: createApplication
      tags: [Applications]
      summary: Create an application
      description: |
        Registers a new tenant.

        Convia assigns the identifier, the status, and the timestamps: a client
        can neither choose an identifier nor backdate a tenant. Creation is not
        idempotent yet, so a repeated request creates a second application.
      x-lifecycle: experimental
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
            example:
              name: Workspace Town
      responses:
        "201":
          description: The application was created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listApplications
      tags: [Applications]
      summary: List applications
      description: |
        Returns one page of applications, newest first. Deleted applications are
        not returned.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: One page of applications.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}:
    get:
      operationId: getApplication
      tags: [Applications]
      summary: Retrieve an application
      description: |
        Returns one application.

        An identifier that does not exist, is malformed, or belongs to a deleted
        application is reported the same way, so that probing reveals nothing
        about which applications exist.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          description: The requested application.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: renameApplication
      tags: [Applications]
      summary: Rename an application
      description: |
        Changes the display name of an application.

        The name is display metadata and carries no security meaning, so
        renaming an application never changes what it may do.

        Send the `ETag` from a previous read as `If-Match` to make the change
        conditional: the rename then applies only while nothing else has
        changed, and is rejected with `412` otherwise. The header is optional,
        and without it the last write wins.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameApplicationRequest"
            example:
              name: Workspace Town
      responses:
        "200":
          description: The renamed application.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteApplication
      tags: [Applications]
      summary: Delete an application
      description: |
        Removes an application from the API surface.

        The record is retained for the erasure window rather than destroyed, so
        the deletion stays recoverable. Deleting an application that is already
        deleted succeeds, because a repeated delete must not fail.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "204":
          description: The application was deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/suspend:
    post:
      operationId: suspendApplication
      tags: [Applications]
      summary: Suspend an application
      description: |
        Withdraws access without losing data. A suspended application is not
        served, and its data is retained until it is activated again.

        Suspending an already-suspended application changes nothing and reports
        success, so a repeated request is safe.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          $ref: "#/components/responses/ApplicationState"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/activate:
    post:
      operationId: activateApplication
      tags: [Applications]
      summary: Activate an application
      description: |
        Restores a suspended application to normal service.

        Activating an already-active application changes nothing and reports
        success, so a repeated request is safe.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      responses:
        "200":
          $ref: "#/components/responses/ApplicationState"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/users:
    post:
      operationId: resolveUser
      tags: [Users]
      summary: Resolve an application's person into a Convia user
      description: |
        Maps one of the application's people to a Convia user, creating the
        mapping the first time it is seen.

        The operation is idempotent by `external_subject`: calling it again
        returns the same user, so an application can call it before every
        session without tracking whether it has called it before. An existing
        user is returned unchanged, so a routine resolve never overwrites a
        display name or metadata that was corrected elsewhere.

        The response is `201` when the mapping was created and `200` when it
        already existed.

        A subject whose user was deleted stays reserved until erasure, so
        resolving it again is refused with `409` rather than reviving the user.
        A suspended user is returned normally, with its status, so the
        application can see the state it set.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveUserRequest"
            example:
              external_subject: customer-42
              display_name: Ada Lovelace
              metadata:
                plan: pro
      responses:
        "200":
          description: The mapping already existed.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "201":
          description: The mapping was created.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listUsers
      tags: [Users]
      summary: List an application's users
      description: |
        Returns one page of the application's users, newest first. Deleted
        users are not returned, and no other application's users can appear.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: One page of users.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/users/{user_id}:
    get:
      operationId: getUser
      tags: [Users]
      summary: Retrieve a user
      description: |
        Returns one user of the application.

        An identifier that does not exist, is malformed, belongs to a deleted
        user, or belongs to another application is reported identically, so
        that this endpoint cannot be used to discover which identifiers exist
        elsewhere in Convia.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: The requested user.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateUser
      tags: [Users]
      summary: Update a user's attributes
      description: |
        Changes the attributes the application owns.

        Only the fields present in the body change, so a display name can be
        corrected without restating metadata. Metadata is replaced rather than
        merged: the object sent becomes the stored object, which is how a key
        is removed.

        Send the `ETag` from a previous read as `If-Match` to make the change
        conditional: the update then applies only while nothing else has
        changed, and is rejected with `412` otherwise. The header is optional,
        and without it the last write wins.

        A deleted user cannot be updated and is reported as missing.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              display_name: Ada King
              metadata:
                plan: enterprise
      responses:
        "200":
          description: The updated user.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete a user
      description: |
        Removes a user from the API surface.

        The record is retained for the erasure window rather than destroyed, so
        the deletion stays recoverable. Deleting a user that is already deleted
        succeeds, because a repeated delete must not fail.

        The `external_subject` stays reserved by the deleted user until erasure
        frees it, so resolving that subject again is refused with `409` rather
        than reviving the user.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: The user was deleted.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/users/{user_id}/suspend:
    post:
      operationId: suspendUser
      tags: [Users]
      summary: Suspend a user
      description: |
        Withdraws a user's access without losing data. The record is retained
        until the user is activated again.

        Suspending an already-suspended user changes nothing and reports
        success, so a repeated request is safe.

        Suspension does not free the external subject: resolving it returns the
        suspended user, so the application can see the state it set.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          $ref: "#/components/responses/UserState"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/applications/{application_id}/credentials/{credential_id}:
    get:
      operationId: getCredential
      tags: [Credentials]
      summary: Retrieve a credential
      description: |
        Returns one credential of the application, never its secret.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/CredentialID"
      responses:
        "200":
          description: The requested credential.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: revokeCredential
      tags: [Credentials]
      summary: Revoke a credential
      description: |
        Withdraws a credential immediately.

        There is no grace period: the next request presenting the key fails,
        because every authentication reads the stored record. Revoking an
        already-revoked credential succeeds, because a repeated request must
        not fail.

        The record is retained rather than deleted, so that an operator can
        still see what was issued and when it was withdrawn.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/CredentialID"
      responses:
        "204":
          description: The credential was revoked.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/applications/{application_id}/users/{user_id}/activate:
    post:
      operationId: activateUser
      tags: [Users]
      summary: Activate a user
      description: |
        Restores a suspended user to normal service.

        Activating an already-active user changes nothing and reports success,
        so a repeated request is safe. A deleted user cannot be activated and is
        reported as missing.
      x-lifecycle: experimental
      parameters:
        - $ref: "#/components/parameters/ApplicationID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          $ref: "#/components/responses/UserState"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApplicationKey:
      type: http
      scheme: bearer
      description: |
        An application authenticates with the API key it was issued, presented
        as `Authorization: Bearer cvk_...`.

        The key is opaque: clients must never parse it. The application Convia
        acts for is taken from the key, which is why the authenticated
        endpoints name no application in their paths.

        Every operation additionally requires the scopes listed on it. A valid
        key without the required scope is answered with `403 forbidden`.
  schemas:
...
    CreateApplicationRequest:
      type: object
      description: The accepted body when creating an application.
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          description: |
            Display name. Surrounding whitespace is insignificant, and control
            characters are rejected.
          minLength: 1
          maxLength: 120
...
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: |
        Maximum number of items to return in one page. Values above the maximum
        are rejected with `invalid_request` rather than silently clamped.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 25
...
  headers:
    RequestId:
      description: |
        Correlation identifier of the request. Convia echoes a safe
        client-supplied `X-Request-ID` and otherwise generates one.
      required: true
      schema:
        type: string
        maxLength: 128
...